Pixels with red=42 will be considered doors.
Doors are now added as "template". This means that in the output file
there will be a `$key$` flag that must be replaced with the position of
the key for that door.

The main problem is that the key/door format is not good enough because
keys should not be in the map file. But for benchmarking and testing
this is the easier and more solid choice.